Delete sessions after successful logout instead of before logout
Deleting a session for an itsme authentication before the logout flow is started causes the flow to terminate in the consent app before the user is redirected to the itsme logout service. So when the user attempts to log in again, when the user reaches the itsme server to log in, the pre-existing session is used to automatically log the user in.

See: https://jira.antwerpen.be/browse/AUTH-394?focusedCommentId=227930&page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#comment-227930